Fix user cannot write to home if vols mounted there.
Problem:
When running the image and mounting a volume into /home/user, the
/home/user folder is owned by root, and user cannot write to it.

Cause:
Docker sees that the /home/user directory does not exist when a
volume is mounted there, and so it creates the missing directory.

When Docker does this, it does so as root, and so the directory
is owned by root with 755 permissions.

Fix:
I updated the entrypoint.sh script to only create the home folder
if it does not yet exist, and ensure that the home folder is
owned by user after creating the user account.