Remove 'Starting with UID' message
Although the message is sometimes useful, it creates non-standard output that may confuse
users who are using Node without knowing that it is running inside a Docker container.

Since one of the goals of this image is to allow Node to be ran as if it were installed
natively, I decided to remove this to avoid any confusion.

It is still possible to modify the docker run statement, or exec into an existing docker
container to determine this information.