Remove node user from image
Problem:
The node user id using user id 1000. There are times when this is the user id that
the Docker host is using, and the Docker user may want to run the container under
this id.

With the way that the entrypoint.sh script currently works, it will create a second
user with the same id# if that id is specified as the LOCAL_USER_ID. Although this
works to some degree, the POSIX permissions are all messed up, and the /home/user
folder ends up being owned by the node user. Even though both users have the same
id, the OS is now allowing the "user" user to read the files, causing problems
such as host verification failures, because the "user" user cannot read
~/.ssh/known_hosts

Fix:
I deleted the node user from the image, as it is no longer needed.