Improve entrypoint.sh to adopt existing user
Problem:
If an LOCAL_USER_ID is provided that is already in use, entrypoint.sh will
go ahead and create a new user with the same id. This can cause a bunch
of issues (see last commit).

Fix:
I modified the entrypoint.sh file so that if the LOCAL_USER_ID exists in
the system, then the script simply uses the existing account instead of
creating another user.

This is useful if you want to run the image as root. This might have been
a potential alternative solution to the fix that the previous commit
addressed, if it were not for the fact that user id# is so prevelant.

We ideally want the image to be as portable as possible, and having user
1000 already exist with a home of /home/node, but most other user ids
ending up with a home of /home/user, means that it would not be possible
to have a single script to run this container and mount ~/.ssh into
a common location regardless of the Docker host.