Fix container changes volume owner
Problem:
When the Docker host mounts a volume into /workdir on Linux, the Docker container changes the ownder and permissions on the volume, which also modifies the hosts filesystem. Oops.

Fix:
I removed the chmod and chown. It was only added to prevent permissions errors of the container was started without a workdir volume mounted, and attempted to write to the workdir, which should be a very unlikely situation, and is not really supported anyway.